Pushed common properties from Participant and Instructor to ParticipatingPerson
