+++ app/model/DTO/Participant/ParticipatingPerson.php
@@ -4,30 +4,46 @@ declare(strict_types=1);
 
 namespace Model\DTO\Participant;
 
+use Model\Participant\Payment;
+use Model\Utils\MoneyFactory;
 use Nette\SmartObject;
 
 /**
  * @property-read int $id
+ * @property-read int $personId
  * @property-read string $firstName
  * @property-read string $lastName
  * @property-read string $nickName
  * @property-read string $displayName
+ * @property-read float $payment
+ * @property-read float $repayment
+ * @property-read string $onAccount
  */
 class ParticipatingPerson
 {
     use SmartObject;
 
+    private Payment $paymentObj;
+
     public function __construct(
         private int $id,
+        private int $personId,
         private string $firstName,
         private string $lastName,
         private string|null $nickName = null,
+        Payment $payment,
     ) {
+        $this->paymentObj = $payment;
     }
 
     public function getId(): int
     {
         return $this->id;
+    }
+
+    public function getPersonId(): int
+    {
+        return $this->personId;
     }
 
     public function getFirstName(): string
@@ -48,5 +64,20 @@ class ParticipatingPerson
     public function getDisplayName(): string
     {
         return $this->lastName . ' ' . $this->firstName . ($this->nickName !== null ? '(' . $this->nickName . ')' : '');
+    }
+
+    public function getPayment(): float
+    {
+        return MoneyFactory::toFloat($this->paymentObj->getPayment());
+    }
+
+    public function getRepayment(): float
+    {
+        return MoneyFactory::toFloat($this->paymentObj->getRepayment());
+    }
+
+    public function getOnAccount(): string
+    {
+        return $this->paymentObj->getAccount();
     }
 }
